Improve delete namespace test
This test was causing some exceptions on the server-side (that are
not problematic, just polluting the logs) because it was deleting
the namespace right after an insert, in almost the same time as
the transient indexing is being done.

This was causing us to reject transient indexing operation and log
that exception in server's log. This exception is totally safe and
can result in normal execution of our APIs, but this test was causing
it from time to time.

I have modified the test to wait for transient indexing to finish
and removed the unnecessary sleep after the namespace is deleted, as the
delete namespace operation is visible to the list namespaces operation
immediately.